Add service  unaviable exception, modify delete hostel method in hostel service

+++ AmdarisProject.Core/Services/HostelService.cs
@@ -4,6 +4,7 @@ using AmdarisProject.Core.Interfaces;
 using AmdarisProject.DataAccess.Interfaces;
 using AmdarisProject.Domain;
 using AutoMapper;
+using System.Reflection.Metadata.Ecma335;
 
 namespace AmdarisProject.Core.Services
 {
@@ -63,7 +64,15 @@ namespace AmdarisProject.Core.Services
 
         public async Task DeleteHostelAsync(int id)
         {
+            var hostel = await _repository.GetByIdAsync(id);
+
+            if (hostel is null) throw new NotFoundException("Hostel isn't exists.");
+
             await _repository.DeleteByIdAsync(id);
+
+            hostel = await _repository.GetByIdAsync(id);
+
+            if (hostel is not null) throw new ServiceUnaviableException("Service isn't aviable");
         }
     }
 }
